Fix bug 56 [entities/edit][Version: 17.7] Error Line: 340

+++ inc/Entities_class.php
@@ -337,18 +337,20 @@ class Entities {
                             }
 
                             /* Delete removed entries */
-                            $coveredcountries_keys = array_map('intval', $coveredcountries_keys);
-                            $covctryodelete = EntitiesContractCountries::get_contractcountries('eid='.$this->eid.' AND coid NOT IN ('.implode(', ', $coveredcountries_keys).')', array('returnarray' => true));
+                            if(is_array($coveredcountries_keys)) {
+                                $coveredcountries_keys = array_map('intval', $coveredcountries_keys);
+                                $covctryodelete = EntitiesContractCountries::get_contractcountries('eid='.$this->eid.' AND coid NOT IN ('.implode(', ', $coveredcountries_keys).')', array('returnarray' => true));
 
-                            if(is_array($covctryodelete)) {
-                                foreach($covctryodelete as $covctry) {
-                                    if(!is_object($covctry)) {
-                                        continue;
+                                if(is_array($covctryodelete)) {
+                                    foreach($covctryodelete as $covctry) {
+                                        if(!is_object($covctry)) {
+                                            continue;
+                                        }
+                                        $covctry->delete();
                                     }
-                                    $covctry->delete();
                                 }
+                                unset($coveredcountries_keys, $covctryodelete);
                             }
-                            unset($coveredcountries_keys, $covctryodelete);
                         }
 
                         /* $query = $db->query("SELECT uid FROM ".Tprefix."assignedemployees WHERE isValidator='1' AND eid='".$this->eid."'");
